+++ tw-stock-analysis/backend/app/models/sync_log.py
@@ -3,7 +3,7 @@ ORM Model — 資料同步紀錄
 """
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Integer, String, Text, func
+from sqlalchemy import BigInteger, Integer, DateTime, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.database import Base
@@ -13,7 +13,7 @@ class SyncLog(Base):
     """資料同步紀錄"""
     __tablename__ = "sync_log"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     dataset: Mapped[str] = mapped_column(String(100), nullable=False)
     stock_id: Mapped[str | None] = mapped_column(String(10))
     sync_start: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
